requirements.txt
pandas>=1.3
numpy>=1.21
pyyaml
python-dateutil
pytest
rich
